Rename downloader file to avoid duplicates
