+++ examples/python/open3d_example.py
@@ -1,0 +1,77 @@
+# ----------------------------------------------------------------------------
+# -                        Open3D: www.open3d.org                            -
+# ----------------------------------------------------------------------------
+# The MIT License (MIT)
+#
+# Copyright (c) 2018-2021 www.open3d.org
+#
+# Permission is hereby granted, free of charge, to any person obtaining a copy
+# of this software and associated documentation files (the "Software"), to deal
+# in the Software without restriction, including without limitation the rights
+# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+# copies of the Software, and to permit persons to whom the Software is
+# furnished to do so, subject to the following conditions:
+#
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
+#
+# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
+# IN THE SOFTWARE.
+# ----------------------------------------------------------------------------
+
+# Helpers and monkey patches for ipynb tutorials
+import open3d as o3d
+import numpy as np
+import PIL.Image
+import IPython.display
+import os
+import urllib.request
+import tarfile
+import gzip
+import zipfile
+import shutil
+import sys
+
+# Whenever you import open3d_example, the test data will be downloaded
+# automatically to Open3D/examples/test_data/open3d_downloads. Therefore, make
+# sure to import open3d_example before running the examples.
+# See https://github.com/isl-org/open3d_downloads for details on how to
+# manage the test data files.
+_pwd = os.path.dirname(os.path.realpath(__file__))
+sys.path.append(os.path.join(_pwd, os.pardir, "test_data"))
+from download_utils import download_all_files as _download_all_files
+_download_all_files()
+
+def _relative_path(path):
+    script_path = os.path.realpath(__file__)
+    script_dir = os.path.dirname(script_path)
+    return os.path.join(script_dir, path)
+
+def get_bunny_mesh():
+    bunny_path = _relative_path("../test_data/Bunny.ply")
+    if not os.path.exists(bunny_path):
+        print("downloading bunny mesh")
+        url = "http://graphics.stanford.edu/pub/3Dscanrep/bunny.tar.gz"
+        urllib.request.urlretrieve(url, bunny_path + ".tar.gz")
+        print("extract bunny mesh")
+        with tarfile.open(bunny_path + ".tar.gz") as tar:
+            tar.extractall(path=os.path.dirname(bunny_path))
+        shutil.move(
+            os.path.join(
+                os.path.dirname(bunny_path),
+                "bunny",
+                "reconstruction",
+                "bun_zipper.ply",
+            ),
+            bunny_path,
+        )
+        os.remove(bunny_path + ".tar.gz")
+        shutil.rmtree(os.path.join(os.path.dirname(bunny_path), "bunny"))
+    mesh = o3d.io.read_triangle_mesh(bunny_path)
+    mesh.compute_vertex_normals()
+    return mesh
